pyproject.toml
[project]
name = "pyespdb"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "dash>=4.0.0",
    "dash-bootstrap-components>=2.0.0",
    "plotly>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "pandas>=2.0.0",
    "mysql-connector-python>=8.0.0",
    "pymysql>=1.0.0",
    "oracledb>=2.0.0",
    "anthropic>=0.40.0",
]
